requirements.txt
st-gsheets-connection
pandas
plotly
streamlit-float